transition to DetectionResult
